Split out SDK_VERSION changes from syncing codegen files
