+++ src/version.ts
@@ -1,1 +1,1 @@
-export const VERSION = "120240313.0.1";
+export const VERSION = "120240313.0.2";
